fix: revert @payloadcms/plugin-sentry to 3.30.0
