+++ src/app/(payload)/api/checkin-app/auth/route.ts
@@ -1,6 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server'
 import { isAdminOrSuperAdminOrEventAdmin } from '@/access/isAdminOrSuperAdmin'
 import { getPayload } from '@/payload-config/getPayloadConfig'
+import { cookies } from 'next/headers'
 
 export async function POST(req: NextRequest) {
   try {
@@ -30,6 +31,13 @@ export async function POST(req: NextRequest) {
         { status: 403 },
       )
     }
+
+    const cookieStore = await cookies()
+    cookieStore.set('payload-token', result.token as string, {
+      maxAge: 60 * 60 * 24 * 1, // 1 day
+      httpOnly: true,
+      secure: process.env.NODE_ENV === 'production',
+    })
 
     return NextResponse.json({
       token: result.token,
