+++ src/app/(payload)/api/checkin-app/checkin/[ticket-code]/route.ts
@@ -5,10 +5,10 @@
 // return ticket details
 
 import { NextRequest, NextResponse } from 'next/server'
-import { headers as getHeaders } from 'next/headers'
 import { isAdminOrSuperAdminOrEventAdmin } from '@/access/isAdminOrSuperAdmin'
 import { getPayload } from '@/payload-config/getPayloadConfig'
 import { revalidateTag } from 'next/cache'
+import { checkAuthenticated } from '@/utilities/checkAuthenticated'
 // import { getClientSideURL } from '@/utilities/getURL'
 
 export async function POST(req: NextRequest) {
@@ -16,17 +16,18 @@ export async function POST(req: NextRequest) {
     // Get authorization header
     const payload = await getPayload()
 
-    const headers = await getHeaders()
-    const { user } = await payload.auth({ headers })
+    const authData = await checkAuthenticated()
 
     if (
-      !user ||
+      !authData?.user ||
       !isAdminOrSuperAdminOrEventAdmin({
-        req: { user },
+        req: { user: authData.user },
       })
     ) {
       return NextResponse.json({ error: 'Unauthorized - Invalid admin user' }, { status: 401 })
     }
+
+    const user = authData.user
 
     // Get ticket code from URL parameter
     const ticketCode = req.nextUrl.pathname.split('/').pop()
